+++ integrationtests.py
@@ -481,6 +481,79 @@ class TestRunParallel(unittest.TestCase):
                 self.assertEqual(stats.numCacheEntries(), 2)
 
 
+# Compiler calls with multiple sources files at once, e.g.
+# cl file1.c file2.c
+class TestMultipleSources(unittest.TestCase):
+    def testTwo(self):
+        with cd(os.path.join(ASSETS_DIR, "mutiple-sources")), tempfile.TemporaryDirectory() as tempDir:
+            cache = clcache.Cache(tempDir)
+            customEnv = dict(os.environ, CLCACHE_DIR=tempDir)
+            baseCmd = CLCACHE_CMD + ["/nologo", "/EHsc", "/c"]
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 0)
+                self.assertEqual(stats.numCacheMisses(), 0)
+                self.assertEqual(stats.numCacheEntries(), 0)
+
+            subprocess.check_call(baseCmd + ["fibonacci01.cpp", "fibonacci02.cpp"], env=customEnv)
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 0)
+                self.assertEqual(stats.numCacheMisses(), 2)
+                self.assertEqual(stats.numCacheEntries(), 2)
+
+            subprocess.check_call(baseCmd + ["fibonacci01.cpp", "fibonacci02.cpp"], env=customEnv)
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 2)
+                self.assertEqual(stats.numCacheMisses(), 2)
+                self.assertEqual(stats.numCacheEntries(), 2)
+
+    def testFive(self):
+        with cd(os.path.join(ASSETS_DIR, "mutiple-sources")), tempfile.TemporaryDirectory() as tempDir:
+            cache = clcache.Cache(tempDir)
+            customEnv = dict(os.environ, CLCACHE_DIR=tempDir)
+            baseCmd = CLCACHE_CMD + ["/nologo", "/EHsc", "/c"]
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 0)
+                self.assertEqual(stats.numCacheMisses(), 0)
+                self.assertEqual(stats.numCacheEntries(), 0)
+
+            subprocess.check_call(baseCmd + [
+                "fibonacci01.cpp",
+                "fibonacci02.cpp",
+                "fibonacci03.cpp",
+                "fibonacci04.cpp",
+                "fibonacci05.cpp",
+            ], env=customEnv)
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 0)
+                self.assertEqual(stats.numCacheMisses(), 5)
+                self.assertEqual(stats.numCacheEntries(), 5)
+
+            subprocess.check_call(baseCmd + [
+                "fibonacci01.cpp",
+                "fibonacci02.cpp",
+                "fibonacci03.cpp",
+                "fibonacci04.cpp",
+                "fibonacci05.cpp",
+            ], env=customEnv)
+
+            with cache.statistics as stats:
+                self.assertEqual(stats.numCacheHits(), 5)
+                self.assertEqual(stats.numCacheMisses(), 5)
+                self.assertEqual(stats.numCacheEntries(), 5)
+
+class TestMultipleSourceWithClEnv(unittest.TestCase):
+    def testAppend(self):
+        with cd(os.path.join(ASSETS_DIR)):
+            customEnv = dict(os.environ, _CL_="minimal.cpp")
+            cmd = CLCACHE_CMD + ["/nologo", "/EHsc", "/c"]
+            subprocess.check_call(cmd + ["fibonacci.cpp"], env=customEnv)
+
+
 class TestClearing(unittest.TestCase):
     def _clearCache(self):
         subprocess.check_call(CLCACHE_CMD + ["-C"])
@@ -614,7 +687,6 @@ class TestNoDirectCalls(unittest.TestCase):
             self.assertEqual(subprocess.call(cmd, env=env), 0) # This should hit now
             with cache.statistics as stats:
                 self.assertEqual(stats.numCacheHits(), oldHits + 1)
-
 
     def testHitViaMpSequential(self):
         with cd(os.path.join(ASSETS_DIR, "parallel")), tempfile.TemporaryDirectory() as tempDir: